requirements.txt
python-dotenv
discord.py
aiohttp
slack-bolt
emoji
pytesseract
openai